Migrate methods to parent class to make available for extend for other gateway.
