+++ includes/Gateway/API.php
@@ -138,7 +138,7 @@ class API extends \WooCommerce\Square\API {
 	 *
 	 * @param \WC_Order $order order object
 	 * @since 3.7.0
-	 * @return \WooCommerce\Square\API\Response
+	 * @return \WooCommerce\Square\Gateway\API\Responses\Create_Payment
 	 */
 	public function gift_card_charge( \WC_Order $order ) {
 		$request = new API\Requests\Payments( $this->get_location_id(), $this->client );
@@ -494,6 +494,7 @@ class API extends \WooCommerce\Square\API {
 	 *
 	 * @param array  $payment_ids Array of payment IDs.
 	 * @param string $order_id    Square order ID.
+	 * @return \WooCommerce\Square\Gateway\API\Responses\Create_PayOrder
 	 * @since 3.9.0
 	 */
 	public function pay_order( $payment_ids, $order_id ) {
